test(perf): Make it easier to run subsets
